+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Better error handling on shader initialization
 - Support for custom textures
 - zh_CN translations ([#6](https://github.com/vfx-dev/SwanSong/pull/6)) (Omgise)
+- Camera position and eye brightness uniforms
 
 ### Fixed